1.3-BETA (#8)
This patch is a improvement patch.

I added a alternative syntax to define schematic sets, which will be parsed into the old legacy syntax. Unfortunately i cant really simplify this system due it's high complexity without removing some features.
The new alternative syntax is way more readable for you. But still features all functions of the old known legacy syntax.

I also improved the tab completion when you type a directory. Instead of listing all directories i only list the folders in the current hierarchy. This makes it more easy to navigate though your folders.

Also the need to manually reload the plugin to use recently added schematics was some kind of lazy implementation. I still need to cache the schematics to ensure, that I don't have to search for a fitting schematic every time, but i added a reload of the schematics every minute. Don't worry this wont cause any lags on your server and will be executed in a separate thread. So now every minute we will add your new schematics to the cache. Of course u can still use the current reload schematics command if you are in a hurry.