Fix: rotation npe (#317)
* Add error handling for SchematicBrushReborn settings and modifiers

The modification introduces error handling for invoking mutations in SchematicBrushReborn's brush settings and schematic modifiers. A try-catch block has been added to catch any runtime exceptions that may occur when applying brush settings or schematic modifiers. If an exception occurs, a warning log message is generated, and if the modifier is shiftable, it is shifted. This approach is aimed at preventing application crash due to a faulty setting or modifier, and providing a clear log message to facilitate in problem diagnosis.

* Add initial value assignments for FlipList and RotationList

A value assignment statement has been added to both FlipList and RotationList classes. This change ensures that the classes have an initial value when they are instantiated. This is important to avoid potential null pointer exceptions and to ensure proper class functionality.

* Fix typo in method documentation

The typo in the method comments of 'Build a paste operation' in the BrushPasteImpl.java file was corrected. The misspelled word 'caputure' is now corrected to 'capture'. This change is purely for enhancing the readability of the code and does not affect any functionality.

* Remove redundant include in settings.gradle.kts

The general include(":") directive has been removed from the settings.gradle.kts file as it was not necessary. All other necessary subprojects are still being included individually which makes the build process clearer and avoids possible complications in the future.