Fix: rendering npe (#356)
* Update version number in build.gradle.kts

The version number in the build.gradle.kts file was updated from 2.6.0 to 2.6.1. This minor version update likely includes bug fixes or very small changes that do not substantially alter the functionality of the project.

* Refactor PacketWorker's remove method

The remove method in PacketWorker has been simplified. The comparison of player unique identifiers is now done in a single line, eliminating the repetitive and more complex boolean logic that existed before, making the code cleaner and easier to understand.

* Refactor resolveBlocked method in RenderService class

The resolveBlocked method in RenderService class has been refactored. We removed the process method and added the remove method from worker object, making the code easier to reason about. This also fixes the issue where playerSink object might not have been correctly updated before performing the remove operation.